+++ contracts/factory/src/contract.rs
@@ -1,11 +1,13 @@
-use soroban_sdk::{contract, contractimpl, contractmeta, log, Address, Env, Vec};
+use soroban_sdk::{
+    contract, contractimpl, contractmeta, log, Address, Env, IntoVal, Symbol, Val, Vec,
+};
 
 use crate::{
     error::ContractError,
-    lp_contract,
     storage::{get_admin, get_lp_vec, save_admin, save_lp_vec},
     utils::deploy_lp_contract,
 };
+use phoenix::utils::{LiquidityPoolInitInfo, StakeInitInfo, TokenInitInfo};
 
 // Metadata that is added on to the WASM custom section
 contractmeta!(key = "Description", val = "Phoenix Protocol Factory");
@@ -18,9 +20,7 @@ pub trait FactoryTrait {
 
     fn create_liquidity_pool(
         env: Env,
-        lp_init_info: lp_contract::LiquidityPoolInitInfo,
-        token_init_info: lp_contract::TokenInitInfo,
-        stake_init_info: lp_contract::StakeInitInfo,
+        lp_init_info: LiquidityPoolInitInfo,
     ) -> Result<(), ContractError>;
 
     fn query_pools(env: Env) -> Result<Vec<Address>, ContractError>;
@@ -42,24 +42,29 @@ impl FactoryTrait for Factory {
 
     fn create_liquidity_pool(
         env: Env,
-        lp_init_info: lp_contract::LiquidityPoolInitInfo,
-        token_init_info: lp_contract::TokenInitInfo,
-        stake_init_info: lp_contract::StakeInitInfo,
+        lp_init_info: LiquidityPoolInitInfo,
     ) -> Result<(), ContractError> {
-        validate_token_info(&env, &token_init_info, &stake_init_info)?;
+        validate_token_info(
+            &env,
+            &lp_init_info.token_init_info,
+            &lp_init_info.stake_init_info,
+        )?;
 
         let lp_contract_address = deploy_lp_contract(&env, lp_init_info.lp_wasm_hash);
 
-        lp_contract::Client::new(&env, &lp_contract_address).initialize(
-            &get_admin(&env)?,
-            &lp_init_info.share_token_decimals,
-            &lp_init_info.swap_fee_bps,
-            &lp_init_info.fee_recipient,
-            &lp_init_info.max_allowed_slippage_bps,
-            &lp_init_info.max_allowed_spread_bps,
-            &token_init_info,
-            &stake_init_info,
-        );
+        let init_fn: Symbol = Symbol::new(&env, "initialize");
+        let init_fn_args: Vec<Val> = (
+            lp_init_info.admin,
+            lp_init_info.share_token_decimals,
+            lp_init_info.swap_fee_bps,
+            lp_init_info.fee_recipient,
+            lp_init_info.max_allowed_slippage_bps,
+            lp_init_info.max_allowed_spread_bps,
+            lp_init_info.token_init_info,
+            lp_init_info.stake_init_info,
+        )
+            .into_val(&env);
+        let _res: Val = env.invoke_contract(&lp_contract_address, &init_fn, init_fn_args);
 
         let mut lp_vec = get_lp_vec(&env)?;
 
@@ -84,8 +89,8 @@ impl FactoryTrait for Factory {
 
 fn validate_token_info(
     env: &Env,
-    token_init_info: &lp_contract::TokenInitInfo,
-    stake_init_info: &lp_contract::StakeInitInfo,
+    token_init_info: &TokenInitInfo,
+    stake_init_info: &StakeInitInfo,
 ) -> Result<(), ContractError> {
     if token_init_info.token_a >= token_init_info.token_b {
         log!(env, "token_a must be less than token_b");
@@ -126,13 +131,13 @@ mod tests {
         let token_a = Address::from_contract_id(&contract1);
         let token_b = Address::from_contract_id(&contract2);
 
-        let token_init_info = lp_contract::TokenInitInfo {
+        let token_init_info = TokenInitInfo {
             token_a,
             token_b,
             token_wasm_hash,
         };
 
-        let stake_init_info = lp_contract::StakeInitInfo {
+        let stake_init_info = StakeInitInfo {
             max_distributions: 10,
             min_bond: 10,
             min_reward: 10,
@@ -158,13 +163,13 @@ mod tests {
         let token_a = Address::from_contract_id(&contract1);
         let token_b = Address::from_contract_id(&contract2);
 
-        let token_init_info = lp_contract::TokenInitInfo {
+        let token_init_info = TokenInitInfo {
             token_a,
             token_b,
             token_wasm_hash,
         };
 
-        let stake_init_info = lp_contract::StakeInitInfo {
+        let stake_init_info = StakeInitInfo {
             max_distributions: 10,
             min_bond: 0,
             min_reward: 10,
@@ -190,13 +195,13 @@ mod tests {
         let token_a = Address::from_contract_id(&contract1);
         let token_b = Address::from_contract_id(&contract2);
 
-        let token_init_info = lp_contract::TokenInitInfo {
+        let token_init_info = TokenInitInfo {
             token_a,
             token_b,
             token_wasm_hash,
         };
 
-        let stake_init_info = lp_contract::StakeInitInfo {
+        let stake_init_info = StakeInitInfo {
             max_distributions: 10,
             min_bond: 10,
             min_reward: 0,